Separate VSP extensions in a file
